Extract hardcoded pieces of TaskrouterCallback into listeners
